requirements.txt
RPi.GPIO==0.6.5
python-uinput==0.11.2
smbus==1.1.post2
